Added bulk transcription and ingestion of videos through the URL. Currently Transcription only.
ayyyyy podcast/audio file support is now in. (was already in via cli but now in via the web GUI as well; though its just a placeholder and doesn't currently work/shouldn't.... quite yet. Need to add more code/options for usage)
Also added a 'Batch Processing' subtab, so that you can ingest and review the transcripts for a series / list of videos.
Currently needs to be provided similar to regular input, so no playlist URLs.

Haven't tested the audio transcription